requirements.txt
numpy
pandas
biopython
bitarray
pyyaml
psutil
memory_profiler
snakemake
pytest
filelock
